requirements.txt
# Database og API
supabase==2.3.0
fastapi==0.104.1
uvicorn==0.24.0

# Vector processing
openai==1.3.5
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
python-multipart==0.0.6

# Email processing
python-email-validator==2.1.0.post1
aiosmtplib==3.0.1

# Testing
pytest==7.4.3
httpx==0.25.2